Large Blank Clearing Update - V2.1
This new update features a large QoL feature a lot of people have requested when trying my game!

Summary:
When you click on an empty tile, it clears all the other empty tiles around it, and the ones around those! It stops at a number.

Tested with large boards with various mine quantities and it finishes clearings in an unnoticeable time frame.

NOTE: Clears also clear any flags, this is required as it's just common sense that empty tiles don't have mines on them.